fix(book-ocr): live_ocr marker の過剰 gating で yomitoku 不要 test まで除外されていた問題を修正
tests/integration/test_book_ocr_yomitoku.py のうち、yomitoku バイナリを呼ばずに
完結する 3 件を tests/unit/test_book_ocr_engine.py に移動し、live_ocr marker を外す:

- test_engine_satisfies_protocol: isinstance チェックのみ
- test_run_batch_empty_returns_empty_list: 空入力で early return
- test_resolve_binary_raises_when_missing: 存在しないパスで FileNotFoundError

これにより CI (tests/unit のみ実行) でも上記 3 件のリグレッション検出が回るようになる。
integration 側には実 OCR が必要な test_run_batch_single_page のみ残す。

Closes #22
